requirements.txt
google-api-python-client==2.21.0
pandas==1.3.3
ipython==7.27.0
python-dateutil==2.8.2
isodate==0.6.0
matplotlib==3.4.3
seaborn==0.11.2
nltk==3.6.2
wordcloud==1.8.1
numpy==1.21.2
